move restcontroller to controllers package
