+++ AssociateEvaluationSystem/src/main/java/com/revature/aes/controllers/FormatRestController.java
@@ -10,7 +10,7 @@
  * 		Edward Crader, Jon-Erik Williams 
  ****************************************************************/
 
-package com.revature.aes.restcontroller;
+package com.revature.aes.controllers;
 
 import java.util.List;
