+++ AssociateEvaluationSystem/src/main/java/com/revature/aes/controllers/TagRestController.java
@@ -1,4 +1,4 @@
-package com.revature.aes.restcontroller;
+package com.revature.aes.controllers;
 
 import java.util.List;
